Add new dynamic UI for websites on settings page
Add input and dropdown elements to eventually replace the textarea.
Since the new elements are not functional yet, I left the textarea in
place for the time being. The basic UI is there though, complete with
add and remove features. The form also pre-populates with the default
sites and their corresponding apps, which are pulled in from a JSON
object. The next step is to save the values from the new elements in
local storage to replace the textarea.