Add Discard Time Notification
Shows a notification to the user when she's been idle for at least 5min while
she was tracking some activity using Toggl.

The notification reads something like:

You've been idle for 14m 53s while tracking "working on the landing
page".

And two options are given to the user:

- Discard idle time
- Discard idle time and stop tracking

This notification/feature can be disabled via the new checkbox in the
eettings page, that reads like:

- [ ] Discard idle time notification