Let Toggl's WebApp know that Toggl Button is installed
Toggl's WebApp could offer a better user experience if it knew whether
the Toggl Button is installed or not. e.g., it could stop upselling /
advertising the existance of the Toggl Button to those users who are
already using it. It also provides additional information to draw a
better conclusion out of A/B experiments.

Toggl's WebApp and website can now tell whether the Toggl Button is
installed or not via checking the existence of `document.TogglButton`.