Fix Trello buttons
The new mutationClass broke Trello buttons in some cases, because there was not one reliable class to respond to in MutationObserver (different things happen when opening a card via a link, etc).

Tweaked the observer matching to match against ALL mutations, rather than just the first mutation in the array of mutations. Then specified multiple selectors for Trello checklists.

Closes #738.