Fix origin checking (app.asana.com)
In the case of subdomains and the like, we do some logic to try and
match a content script even if an immediate match isn't available.

The code was changed recently and the test ended up testing vs null
whereas undefined is really what we're after here.

!x seems to be the case that stops jslint from shouting, whereas it
shouts at 'x == null' and '!(x != null)' and the like with current
settings.

Closes issue #781