Todoist: Use name of containing subproject as Toggl project
Todoist displays entries of a subproject below the main project's
entries. This commit adds the ability to use the subproject name as
the Toggl project.

Any depth of project nesting is handled. We determine the name of the
project's parent, grandparent, great-grandparent, and so on up the
project hierarchy. Then we pass an array of these project names to the
createTimerLink function so that each may be tried successively until
one of them matches a Toggl project.

We determine the parent of a Todoist project by inspecting the
indentation of their entries in the sidebar menu.