Fix Bitrix24 Integration
The layout for Bitrix24 has changed since the last commit. This fix leaves legacy support for those who have self-hosted bitrix24 installations past the update support window. In addition, there is a task popover function, which displays in an iframe. The contents of the iframe aren't accessible by the render method, so I added that functionality in the callback. The styles for ".toggl-button" needed to be inserted into the iframe, so I hardcoded them. Not ideal, but it works.