Fixes project selector for Workast 'My tasks' page
The My tasks page has the title in an `<a>` element while other
task pages have the title in a `<span>` element. This changes the
selector from fetching the first `span` to just selecting the
`firstElementChild`.

The result is that the list header is used for the project title
on the 'My tasks' page.