fix(redmine) Prevent positioning in sidebar
Currently the button is rendered in the first `h2` tag found. Depending
on installed plugins in Redmine there are multiple `h2` tags present.
This leads to having the button rendered at the wrong position and using
the wrong title.

This PR makes the selectors more specific by using `#content h2` to
detect `numElem` and button position.