fix: YouTrack integration
When starting a timer from card popup in agile board wrong issue ID was
referenced. I believe that the logic for finding the element containing
the issue ID was querying the whole document resulting in the wrong ID
in most of the cases. This should fix that by finding the Toggle
button's parent element and then finding the child element of that
parent that contains the issue ID.

closes #1509