fix(popup): Fix tag/project dropdown toggle
There are two event handlers configured that are triggering the popdown (by calling the 'toggleDropdown' function):

 1. A focus event configured in the class 'TB_Dialog_field' (parent of the element with id 'toggl-button-(tag|project)-placeholder')
 2. A click event configured in the element with id 'toggl-button-(tag|project)-placeholder'

This commit removes the 'focus' event handler from 'TB_Dialog_field', since it seems that it's no longer necessary.
Another possible solution would be to add a 'e.stopPropagation' on the click event handler.

Closes #1422