Attach a `pomodoro-session` attribute to each time entry in the popup.
A time entry is either a break or a work session. When clicking to start an entry, we need to know if it is a break or work entry so that we can program the Pomodoro behaviour accordingly. A work entry will be stopped, then followed by a new break entry, but a break entry will be stopped when time is out.