fix(fitlab): Fix tracking button not showing on issue page with new look
Please remember the Contributing Guidelines ❤️

🌟 What does this PR do?
Problems:
The button was not showing in the new Gitlab look on issue detail. The new look was optional up until now.

Solution
Changed selector for issue button. Split issue and merge request title getters since the original one does not work for both of them.
Changed the getter for project since it is broken on the new issue look page (weird or broken breadcrumbs)
The ID getter now queries the body element which has the id in an attribute.
Tags selector now check for duplicate tags.

🐛 Recommendations for testing
Issue page: See if the button appears on the issue page with the new look.
Both (Issue and merge request page): Check if all tags that the issue has assigned are available for tracking. Check if the title is correct. Check the project name